// src/api/models/dtos/Response/RecipeRes/RecipeRes.ts
import { Nutrition_InfoEntity } from '../../../entities/Nutrition_InfoEntity/Nutrition_InfoEntity';
import { OccasionEntity } from '../../../entities/OccasionEntity/OccasionEntity';
import { RecipeEntity } from '../../../entities/RecipeEntity/RecipeEntity';
import { AuthorRes } from '../AuthorRes/AuthorRes';
import { CommentRes } from '../CommentRes/CommentRes';
import { DirectionRes } from '../DirectionRes/DirectionRes';
import { IngredientRes } from '../IngredientRes/IngredientRes';
import { RelatedRecipeRes } from '../RelatedRecipeRes/RelatedRecipeRes';

export type RecipeRes = {
  id: RecipeEntity['id'];
  name?: RecipeEntity['name'];
  is_private: RecipeEntity['is_private'];
  rating?: RecipeEntity['rating'];
  totalTime?: RecipeEntity['totalTime'];
  serving_size: RecipeEntity['serving_size'];
  introduction?: RecipeEntity['introduction'];
  author_note?: RecipeEntity['author_note'];
  image?: RecipeEntity['image'];
  author: AuthorRes;
  ingredients: IngredientRes[];
  nutrition_info: Nutrition_InfoEntity;
  directions: DirectionRes[];
  comments: CommentRes[];
  createAt?: Date;
  occasions?: OccasionEntity[];
  relatedRecipes?: RelatedRecipeRes[];
  isDeleted?: boolean;
};